improving the mimicry behavior (revisions will be made)
This update improves the mimicry behavior, addressing a needed change (imo). previously the mimic's voice lines often gave away its true nature, as the speech patterns didn't make sense and were too random, and thus making it so fucking obvious that it was a mimic. This change should provide a more realistic and immersive/realistic experience.